Bugfixes and major cleanup
- the major change is a functional GenerateModel class that
takes in the job settings and three dataframes specifying paramter
inputs, interaction strengths, and parameter sets.
- The interaction strength code had some suspicious lines which
  have been cleaned up, with more checks to ensure that withRules
  and withoutRules are updated as nodes are assigned parameter inputs
  or self activation loops
- Added some error handling in the post processing functions. If post
  processing functions are called without the correct simulation
  files, the code now suggests what should be fixed in the config file
- Started adding new class documentation
This commit is tested to some extent, but the output needs to be
verified more concretely. Furthermore, while the heaviside function
option seems to be functional, the concepts of parameter inputs and
interaction strengths don't translate very well for heaviside
functions. I am going to document this for now but let someone
else thing about the parameters that have to be added to handle this.